Completely overhauled world generation architecture
Deprecated the old world generation script with a modular generation system.

It functions in stages. Starting by generating a base of land mass (land and water) then applying modifications to that world.

Each routine is a different stage object. It's easy to add more or modify stages.

Old world generation script still exists for now, and i created a wrapper for the new system so that the old script can be ran as a stage in the new system. 

Old system constants will be moved to a GenerationCompanion object.

Debug
====
When in debug mode, Numpad plus is bound to increase the 3d height of the camera, so you can get a better overview of the world.

F3 also adds the test world generation configuration window to the game.

+++ core/src/com/shinkson47/SplashX6/game/world/Voronoi.java
@@ -31,7 +31,7 @@ public class Voronoi {
         for (int x = 0; x < size; x++) {
             for (int y = 0; y < size; y++) {
                 n = 0;
-                for (byte i = 0; i < cells; i++) {
+                for (int i = 0; i < cells; i++) {
                     if (distance(px[i], x, py[i], y) < distance(px[n], x, py[n], y)) {
                         n = i;
                     }
